requirements.txt
wyoming>=1.8.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
esphome
faster-whisper>=0.10.0
piper-tts>=1.2.0
fastapi>=0.100.0
uvicorn>=0.24.0
